feat: transaction confirmation state

+++ bamwallet/Model/NetworkSettings.cs
@@ -9,7 +9,7 @@ namespace BAMWallet.Model
         public string WalletEndpoint { get; set; }
         public string RemoteNode { get; set; }
         public string RemoteNodePubKey { get; set; }
-        public uint NumberOfConfirmations { get; set; }
+        public ulong NumberOfConfirmations { get; set; }
         public bool RunSilently { get; set; }
         public bool Staking { get; set; }
         public string StakeListening { get; set; }
